requirements.txt
appnope==0.1.0
-e git+https://github.com/danilobellini/audiolazy@2a2fabadf955bd0c8b39334b10dfe68c6d85edac#egg=audiolazy
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cycler==0.10.0
dask==0.8.0
decorator==4.0.9
eventlet==0.18.4
Flask==0.10.1
Flask-Bower==1.2.1
Flask-SocketIO==2.1
functools32==3.2.3.post2
gnureadline==6.3.3
greenlet==0.4.9
ipython==4.1.2
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jsonpickle==0.9.3
jsonschema==2.5.1
MarkupSafe==0.23
matplotlib==1.5.1
music21==2.2.1
networkx==1.11
nose==1.3.7
numpy==1.10.4
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
pylab==0.1.3
pyOSC==0.3.5b5294
pyparsing==2.1.0
python-dateutil==2.5.0
python-engineio==0.8.8
python-socketio==1.0
pytz==2015.7
pyzmq==15.2.0
Rx==1.2.6
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
seaborn==0.7.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sympy==0.7.6.1
Theano==0.7.0
-e git+https://github.com/JonathanRaiman/theano_lstm@cde036da62aeed1b661a76f55f13db7a6fe52e63#egg=theano_lstm
toolz==0.7.4
tornado==4.3
traitlets==4.1.0
Werkzeug==0.11.4
